pyproject.toml
[project]
name = "thoth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.154.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pre-commit>=4.0.1",
    "psycopg2>=2.9.10",
    "send2trash>=1.8.3",
    "watchdog>=6.0.0",
]
